[MIN-133] Implement a macro for mocks
