+++ Cargo.toml
@@ -28,6 +28,7 @@ members = [
     'primitives',
     'service',
     'runtime',
+    'helper',
     'pallets/liquidity-pools',
     'pallets/m-tokens',
     'pallets/minterest-model',
