+++ pallets/dex/Cargo.toml
@@ -16,6 +16,7 @@ orml-traits = { path = "../../orml/traits", default-features = false }
 
 minterest-primitives = { path = "../../primitives", default-features = false }
 pallet-traits = { path = "../traits", default-features = false }
+helper = { path = "../../helper" }
 
 [dev-dependencies]
 sp-core = "3.0.0"
@@ -40,4 +41,5 @@ std = [
 
 	"minterest-primitives/std",
 	"pallet-traits/std",
+	"helper/std",
 ]
