+++ pallets/integration-tests/src/lib.rs
@@ -29,12 +29,15 @@ mod tests {
 	use minterest_protocol::Error as MinterestProtocolError;
 	use pallet_traits::{PoolsManager, PriceProvider};
 	use sp_std::cell::RefCell;
+	use helper::impl_system_config;
 
 	mod controller_tests;
 	mod liquidity_pools_tests;
 	mod minterest_model_tests;
 	mod minterest_protocol_tests;
 	mod scenario_tests;
+
+	pub type AccountId = u64;
 
 	type UncheckedExtrinsic = frame_system::mocking::MockUncheckedExtrinsic<Test>;
 	type Block = frame_system::mocking::MockBlock<Test>;
@@ -59,37 +62,7 @@ mod tests {
 		}
 	);
 
-	parameter_types! {
-		pub const BlockHashCount: u64 = 250;
-		pub const SS58Prefix: u8 = 42;
-	}
-
-	pub type AccountId = u64;
-
-	impl system::Config for Test {
-		type BaseCallFilter = ();
-		type BlockWeights = ();
-		type BlockLength = ();
-		type DbWeight = ();
-		type Origin = Origin;
-		type Call = Call;
-		type Index = u64;
-		type BlockNumber = u64;
-		type Hash = H256;
-		type Hashing = BlakeTwo256;
-		type AccountId = u64;
-		type Lookup = IdentityLookup<Self::AccountId>;
-		type Header = Header;
-		type Event = Event;
-		type BlockHashCount = BlockHashCount;
-		type Version = ();
-		type PalletInfo = PalletInfo;
-		type AccountData = ();
-		type OnNewAccount = ();
-		type OnKilledAccount = ();
-		type SystemWeightInfo = ();
-		type SS58Prefix = SS58Prefix;
-	}
+	impl_system_config!(Test);
 
 	type Amount = i128;
