+++ pallets/liquidity-pools/Cargo.toml
@@ -23,6 +23,7 @@ orml-tokens = { path = "../../orml/tokens", default-features = false }
 
 minterest-primitives = { path = "../../primitives", default-features = false }
 pallet-traits = { path = "../traits", default-features = false }
+helper = { path = "../../helper" }
 
 [dev-dependencies]
 sp-core = { version = "3.0.0" }
@@ -49,4 +50,5 @@ std = [
 
     "minterest-primitives/std",
     "pallet-traits/std",
+    "helper/std",
 ]
