+++ pallets/m-tokens/Cargo.toml
@@ -22,6 +22,7 @@ orml-utilities = { path = "../../orml/utilities", default-features = false }
 orml-tokens = { path = "../../orml/tokens", default-features = false }
 
 minterest-primitives = { path = "../../primitives", default-features = false }
+helper = { path = "../../helper" }
 
 [dev-dependencies]
 sp-core = { version = "3.0.0", default-features = false }
@@ -45,4 +46,5 @@ std = [
     "orml-tokens/std",
 
     "minterest-primitives/std",
+    "helper/std",
 ]
