+++ pallets/minterest-model/Cargo.toml
@@ -18,6 +18,7 @@ orml-utilities = { path = "../../orml/utilities", default-features = false }
 
 minterest-primitives = { path = "../../primitives", default-features = false }
 liquidity-pools = { path = "../liquidity-pools", default-features = false }
+helper = { path = "../../helper" }
 
 [dev-dependencies]
 sp-core = { version = '3.0.0' }
@@ -44,4 +45,5 @@ std = [
 
     "minterest-primitives/std",
     "liquidity-pools/std",
+    "helper/std",
 ]
