+++ pallets/mnt-token/Cargo.toml
@@ -18,6 +18,7 @@ sp-arithmetic = { version = "3.0.0", default-features = false }
 
 minterest-primitives = { path = "../../primitives", default-features = false }
 pallet-traits = { path = "../traits", default-features = false }
+helper = { path = "../../helper" }
 
 [dev-dependencies]
 sp-core = "3.0.0"
@@ -45,4 +46,5 @@ std = [
 
     "pallet-traits/std",
     "minterest-primitives/std",
+	"helper/std",
 ]
