+++ pallets/risk-manager/Cargo.toml
@@ -26,6 +26,7 @@ minterest-protocol = { path = "../minterest-protocol", default-features = false 
 liquidation-pools = { path = "../liquidation-pools", default-features = false }
 pallet-traits = { path = "../traits", default-features = false }
 controller = { path = "../controller", default-features = false }
+helper = { path = "../../helper" }
 
 [dev-dependencies]
 sp-core = { version = '3.0.0' }
@@ -59,4 +60,5 @@ std = [
     "minterest-protocol/std",
     "liquidation-pools/std",
     "controller/std",
+    "helper/std",
 ]
